Linear-time multiplexing for documents over websocket
Reduces the amount of work needed to ignore irrelevant messages due to
documentName mux/demux from O(n^2) to O(n).

When just a few documents are muxed over the same websocket, the extra
work done by the `on('message')` handler in each Connection is not
noticable. However, when a single websocket carries messages for dozens
or hundreds of documents, the work done by each Connection to discover
the documentName becomes exponential, causing slowdown and potential
crashes.